% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_weighted_var_imp}
\alias{get_weighted_var_imp}
\title{Obtain weighted variable importances}
\usage{
get_weighted_var_imp(num_vars, BIC, sum_trees)
}
\description{
Obtain weighted variable importances
}
