% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Obtain draws from gibbs sampler}
\usage{
gibbs_sampler(overall_sum_trees, overall_sum_mat, y, BIC_weights, num_iter,
  burnin, num_obs, num_test_obs, a, sigma, mu_mu, nu, lambda, resids,
  test_data)
}
\description{
Obtain draws from gibbs sampler
}
