% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampler_ITE_no_update}
\alias{gibbs_sampler_ITE_no_update}
\title{Obtain draws from gibbs sampler}
\usage{
gibbs_sampler_ITE_no_update(overall_sum_trees, overall_sum_mat, y,
  BIC_weights, num_iter, burnin, num_obs, num_test_obs, a, sigma, mu_mu,
  nu, lambda, resids, all_treated_data, all_control_data)
}
\description{
Obtain draws from gibbs sampler
}
