% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampler_no_update_new_inits}
\alias{gibbs_sampler_no_update_new_inits}
\title{Obtain draws from gibbs sampler}
\usage{
gibbs_sampler_no_update_new_inits(overall_sum_trees, overall_sum_mat, y,
  BIC_weights, num_iter, burnin, num_obs, num_test_obs, a, sigma, mu_mu,
  nu, lambda, resids, test_data, new_pred_list)
}
\description{
Obtain draws from gibbs sampler
}
