% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITEs_bartBMA.R
\name{ITEs_bartBMA}
\alias{ITEs_bartBMA}
\title{ITE Predictions (in-sample) using bartBMA and the method described by Hill (2011)}
\usage{
ITEs_bartBMA(x_covariates, z_train, y_train, a1 = 3, nu1 = 3,
  sigquant1 = 0.9, c1 = 1000, pen1 = 12, num_cp1 = 20,
  x.test1 = matrix(0, 0, 0), num_rounds1 = 5, alpha1 = 0.95,
  beta1 = 1, split_rule_node1 = 0, gridpoint1 = 0,
  maxOWsize1 = 100, num_splits1 = 5, gridsize1 = 10,
  zero_split1 = 1, only_max_num_trees1 = 1,
  min_num_obs_for_split1 = 2, min_num_obs_after_split1 = 2)
}
\arguments{
\item{x_covariates}{Covaraite matrix for training bartBMA.}

\item{z_train}{treatment vector for traiing bartBMA.}

\item{y_train}{outcome vector for training bartBMA.}

\item{a}{This is a parameter that influences the variance of terminal node parameter values. Default value a=3.}

\item{nu}{This is a hyperparameter in the distribution of the variance of the error term. THe inverse of the variance is distributed as Gamma (nu/2, nu*lambda/2). Default value nu=3.}

\item{sigquant}{??}

\item{c}{This determines the size of Occam's Window}

\item{pen}{This is a parameter used by the Pruned Exact Linear Time Algorithm when finding changepoints. Default value pen=12.}

\item{num_cp}{This is a number between 0 and 100 that determines the proportion of changepoints proposed by the changepoint detection algorithm to keep when growing trees. Default num_cp=20.}

\item{x.test}{Test data covariate matrix. Default x.test=matrix(0.0,0,0).}

\item{num_rounds}{Number of trees. (Maximum number of trees in a sum-of-tree model). Default num_rounds=5.}

\item{alpha}{Parameter in prior probability of tree node splitting. Default alpha=0.95}

\item{beta}{Parameter in prior probability of tree node splitting. Default beta=1}

\item{split_rule_node}{Binary variable. If equals 1, then find a new set of potential splitting points via a changepoint algorithm after adding each split to a tree. If equals zero, use the same set of potential split points for all splits in a tree. Default split_rule_node=0.}

\item{gridpoint}{Binary variable. If equals 1, then a grid search changepoint detection algorithm will be used. If equals 0, then the Pruned Exact Linear Time (PELT) changepoint detection algorithm will be used (Killick et al. 2012). Default gridpoint=0.}

\item{maxOWsize}{Maximum number of models to keep in Occam's window. Default maxOWsize=100.}

\item{num_splits}{Maximum number of splits in a tree}

\item{gridsize}{This integer determines the size of the grid across which to search if gridpoint=1 when finding changepoints for constructing trees.}

\item{zero_split}{Binary variable. If equals 1, then zero split trees can be included in a sum-of-trees model. If equals zero, then only trees with at least one split can be included in a sum-of-trees model.}

\item{only_max_num_trees}{Binary variable. If equals 1, then only sum-of-trees models containing the maximum number of trees, num_rounds, are selected. If equals 0, then sum-of-trees models containing less than num_rounds trees can be selected. The default is only_max_num_trees=1.}

\item{min_num_obs_for_split}{This integer determines the minimum number of observations in a (parent) tree node for the algorithm to consider potential splits of the node.}

\item{min_num_obs_after_split}{This integer determines the minimum number of observations in a child node resulting from a split in order for a split to occur. If the left or right chikd node has less than this number of observations, then the split can not occur.}
}
\value{
A vector of Individual Treatment Effect Estimates.
}
\description{
This function produces ITE Predictions (in-sample) using bartBMA and the method described by Hill (2011).
}
