% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_vars_lin_alg_parallel_outsamp}
\alias{mean_vars_lin_alg_parallel_outsamp}
\title{Obtain BARTBMA predictions}
\usage{
mean_vars_lin_alg_parallel_outsamp(overall_sum_trees, overall_sum_mat, y,
  BIC_weights, num_iter, burnin, num_obs, num_test_obs, a, sigma, mu_mu,
  nu, lambda, test_data, num_cores)
}
\description{
Obtain BARTBMA predictions
}
