% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pred_ints_chol_attempt_outsamp}
\alias{pred_ints_chol_attempt_outsamp}
\title{Obtain BARTBMA predictions}
\usage{
pred_ints_chol_attempt_outsamp(overall_sum_trees, overall_sum_mat, y,
  BIC_weights, num_iter, burnin, num_obs, num_test_obs, a, sigma, mu_mu,
  nu, lambda, test_data, lower_prob, upper_prob)
}
\description{
Obtain BARTBMA predictions
}
