% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ints_from_rmixt_parallel.R
\name{pred_ints_from_rmixt_parallel}
\alias{pred_ints_from_rmixt_parallel}
\title{Prediction intervals for bart-bma output obtained using rmixt function from the package mvnfast}
\usage{
pred_ints_from_rmixt_parallel(object, num_iter, burnin, l_quant, u_quant,
  newdata = NULL, update_resids = 1, num_cores = 1)
}
\arguments{
\item{num_iter}{Total number of iterations of the Gibbs sampler (including burn-in).}

\item{burnin}{Number of burn-on iterations of the Gibbs sampler.}

\item{newdata}{Test data for which predictions are to be produced. Default = NULL. If NULL, then produces prediction intervals for training data if no test data was used in producing the bartBMA object, or produces prediction intervals for the original test data if test data was used in producing the bartBMA object.}

\item{update_resids}{Option for whether to update the partial residuals in the gibbs sampler. If equal to 1, updates partial residuals, if equal to zero, does not update partial residuals. The defaullt setting is to update the partial residua;s.}
}
\value{
The output is a list of length one. The one element in this list is a vector of prediction intervals???
}
\description{
This function produces prediction intervals for bart-bma output obtained using rmixt function from the package mvnfast.
}
