% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pred_ints_ITE_CATT_insamp_par}
\alias{pred_ints_ITE_CATT_insamp_par}
\title{Obtain BARTBMA predictions}
\usage{
pred_ints_ITE_CATT_insamp_par(overall_sum_trees, overall_sum_mat, y,
  BIC_weights, num_obs, a, sigma, mu_mu, nu, lambda, lower_prob,
  upper_prob, num_cores, root_alg_precision, training_data, ztrain)
}
\description{
Obtain BARTBMA predictions
}
