% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pred_ints_lin_alg_insamp}
\alias{pred_ints_lin_alg_insamp}
\title{Obtain BARTBMA predictions}
\usage{
pred_ints_lin_alg_insamp(overall_sum_trees, overall_sum_mat, y,
  BIC_weights, num_iter, burnin, num_obs, a, sigma, mu_mu, nu, lambda,
  lower_prob, upper_prob)
}
\description{
Obtain BARTBMA predictions
}
