% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_means_new_initials.R
\name{pred_means_bbma_new_initials}
\alias{pred_means_bbma_new_initials}
\title{Predictions for bart-bma output obtained from a Gibbs sampler}
\usage{
pred_means_bbma_new_initials(object, num_iter, burnin, newdata = NULL,
  update_resids = 1, trainingdata)
}
\arguments{
\item{object}{bartBMA object obtained from function bartBMA}

\item{num_iter}{Total number of iterations of the Gibbs sampler (including burn-in).}

\item{burnin}{Number of burn-on iterations of the Gibbs sampler.}

\item{newdata}{Test data for which predictions are to be produced. Default = NULL. If NULL, then produces prediction intervals for training data if no test data was used in producing the bartBMA object, or produces prediction intervals for the original test data if test data was used in producing the bartBMA object.}

\item{update_resids}{Option for whether to update the partial residuals in the gibbs sampler. If equal to 1, updates partial residuals, if equal to zero, does not update partial residuals. The defaullt setting is to update the partial residua;s.}
}
\value{
The output is a list of length one. The one element in this list is a vector of prediction intervals???
}
\description{
This function produces predictions from BART-BMA by post-hoc Gibbs-sampling from the full conditionals of the terminal node parameters and the variance of the error term. See Hernandez et al. (2018) Appendix D for details.
}
