% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_probit_bartBMA.R
\name{predict_probit_bartBMA}
\alias{predict_probit_bartBMA}
\title{Predictions for a new dataset using an existing probit_bartBMA object}
\usage{
predict_probit_bartBMA(object, newdata)
}
\arguments{
\item{object}{A probit_bartBMA object obtained using the probit_bartBMA function.}

\item{newdata}{Covariate matrix for new dataset.}
}
\value{
A vector of predictions for the new dataset.
}
\description{
This function produces predictions for a new dataset using a previously obtained bartBMA object.
}
