% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{preds_bbma_lin_alg_insamp}
\alias{preds_bbma_lin_alg_insamp}
\title{Obtain BARTBMA predictions}
\usage{
preds_bbma_lin_alg_insamp(overall_sum_trees, overall_sum_mat, y,
  BIC_weights, num_iter, burnin, num_obs, a, sigma, mu_mu, nu, lambda)
}
\description{
Obtain BARTBMA predictions
}
