% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{preds_bbma_lin_alg_outsamp}
\alias{preds_bbma_lin_alg_outsamp}
\title{Obtain BARTBMA predictions}
\usage{
preds_bbma_lin_alg_outsamp(overall_sum_trees, overall_sum_mat, y,
  BIC_weights, num_iter, burnin, num_obs, num_test_obs, a, sigma, mu_mu,
  nu, lambda, test_data)
}
\description{
Obtain BARTBMA predictions
}
