% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpScores.R
\name{varImpScores}
\alias{varImpScores}
\title{Variable importances as defined by Hernandez et al. (2018)}
\usage{
varImpScores(object, ...)
}
\arguments{
\item{object}{A bartBMA object obtained using the barBMA function.}
}
\value{
A vector of variable importances. The variables are ordered in the same order that they occur in columns of the input covariate matrix used to obtain the input bartBMA object.
}
\description{
This measure defines the importance of a variable as the model-probability weighted sum of the number of splits on the variable of interest, divided by the sum over all variables of such weighted counts of splits.
}
