% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varIncProb.R
\name{varIncProb}
\alias{varIncProb}
\title{Variable inclusion probabilities as defined by Linero (2018)}
\usage{
varIncProb(object, ...)
}
\arguments{
\item{object}{A bartBMA object obtained using the barBMA function.}
}
\value{
A vector of posterior inclusion probabilities. The variables are ordered in the same order that they occur in columns of the input covariate matrix used to obtain the input bartBMA object.
}
\description{
This measure defines the posterior inclusion probability of a variable as the model-probability weighted sum of indicator variables for whether the variable was used in any splitting rules in any of the trees in the sum-of-tree model.
}
